+++ richdoc/richdoc-cli/src/richdoc_cli/publish/__init__.py
@@ -1,4 +1,0 @@
-"""Remote publishing integrations for richdoc HTML.
-
-Currently only Confluence Cloud is supported.
-"""